+++ src/Monster.cpp
@@ -4,7 +4,7 @@ Monster::Monster(std::shared_ptr<b2World>& world, sf::Vector2f pos): m_pos(pos)
 {
 	m_MonsterTex.loadFromFile("Monster.png");
 	m_MonsterImg.setTexture(m_MonsterTex);
-	m_MonsterImg.setOrigin(m_MonsterTex.getSize().x / 2, m_MonsterTex.getSize().y / 2);
+	m_MonsterImg.setOrigin(float(m_MonsterTex.getSize().x / 2), float(m_MonsterTex.getSize().y / 2));
 
 	b2BodyDef BodyDef;
 	BodyDef.position = b2Vec2(100, m_pos.y - 70);
